fix bugs and add message

+++ server/utils/Utils.js
@@ -160,16 +160,22 @@ const updateUsersInWaitingRoom = async (room_id, user_id, action) => {
 };
 
 async function deleteRoom(room_id, user_id) {
-  const res = await admin.deleteRoom(room_id, user_id);
-  const host_user_id = rooms[room_id].find(
-    (participant) => participant.role === HOST,
-  ).user_id;
-  socket
-    .to(user_sockets[host_user_id])
-    .emit("UPDATE_PARTICIPANTS_IN_WAITING_ROOM", { type: "delete" });
-  socket.emit("UPDATE_USER_ROOMS", { type: "delete" });
-  socket.leave(room_id);
-  return res;
+  try {
+    const res = await admin.deleteRoom(room_id, user_id);
+    if (rooms[room_id]) {
+      const host_user_id = rooms[room_id]?.find(
+        (participant) => participant.role === HOST,
+      ).user_id;
+      socket
+        .to(user_sockets[host_user_id])
+        .emit("UPDATE_PARTICIPANTS_IN_WAITING_ROOM", { type: "delete" });
+    }
+    socket.emit("UPDATE_USER_ROOMS", { type: "delete" });
+    socket.leave(room_id);
+    return res;
+  } catch (error) {
+    console.log("error", error);
+  }
 }
 
 async function createMeeting(
